application_controllerの記述をconcernsディレクトリに記述
